# docs/swagger/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "description": "Creates a new user account with the provided details.",
        "operationId": "AuthController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_name": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_name": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict. Username already exists or Password is required or Role is invalid."
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticates a user and returns a JWT token.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_name": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user has been successfully authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_name": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Invalid Username or Password."
          }
        },
        "summary": "User login",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/verify": {
      "post": {
        "description": "Verifies the JWT token from cookies and returns the payload if valid.",
        "operationId": "AuthController_verify",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The token is valid.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sub": {
                      "type": "number"
                    },
                    "user_name": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Invalid or missing token."
          }
        },
        "summary": "Verify JWT token",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "ARAPL RaaS GTP API",
    "description": "\n      Goods To Person (GTP) API for ARAPL RaaS (Robotics as a Service)\n\n      This API provides comprehensive GTP capabilities including:\n      - Order management and file upload processing\n      - Moveops location management\n      - Station management and Pallet Slot assignments\n      - Inventory tracking and bulk operations\n      \n      All endpoints include proper validation, error handling, and support file uploads where applicable.\n    ",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "API Key for authentication"
      }
    },
...
  }
}